build.gradle:   compile for right version of java   shadow to include other jar
buildFile
  fixed typo

StateCounterCommandEncoder.java
  BasicCommandEncoder assumes it is being added to a GameModule. In this
case the Buildable b is an Extension, not a GameModule. So have
to hide BasicCommandEncoder.addTo(), which also means hiding
ASLCommandEncoder.addTo(). I think the code executed by the real
ASLCommandEncoder() in the vasl GameModule will do everything, and
these don't need to be called a second time.